Connect-DB
connected mongoDB
fixed pizza css(select box)
update pizza js
