+++ server/server.js
@@ -31,7 +31,25 @@ app.use(bodyParser.json());
 app.use(bodyParser.urlencoded({ extended: true }));
 
 app.get("/api/loadPizzaData", async (req, res) => {
-    const pizzaList = await db.collection("PizzaData").find().toArray();
+    const pizzaType = req.query.type || "전체";
+    const pizzaSelectedSort = parseInt(req.query.sort) || 2;
+    const currentPage = parseInt(req.query.currentPage) || 1;
+    const perPage = 2; // 한 페이지 당 보여질 피자 개수
+
+    let findOption = (pizzaType === "전체") ? {} : {kind:`${pizzaType}`}
+
+    let sortOption = {};
+    if(pizzaSelectedSort === 1) sortOption = {priceL: 1};
+    else if(pizzaSelectedSort === 0) sortOption = {priceL: -1};
+
+    const pizzaList = await db
+        .collection("PizzaData")
+        .find(findOption)
+        .sort(sortOption)
+        .skip((currentPage - 1) * perPage)
+        .limit(perPage)
+        .toArray();
+
     const result = pizzaList.map((pizza) => ({
         name: pizza.name,
         imgPath: pizza.imgPath,
@@ -48,67 +66,15 @@ app.get("/api/loadPizzaData", async (req, res) => {
     return res.json(result);
 })
 
-//DB 조회가 완료된 후 send 해야하므로 비동기 처리 필수
-//비동기 처리하지 않으면 빈배열만 보내게 됨
-app.get("/Stats/loadData", async (req, res) => {
-    let result = {};
-    let idx = 0;
+app.get("/api/PizzaDataCount", async (req, res) => {
+    const pizzaType = req.query.type || "전체";
+    let findOption = (pizzaType === "전체") ? {} : {kind:`${pizzaType}`}
 
-    const pizzaName = [];
-    await db.collection("pizza").find().sort({"name": 1}).forEach((r) => {
-        pizzaName.push(r.name);
-    });
+    const count = await db
+        .collection("PizzaData")
+        .countDocuments(findOption); // 총 데이터 개수
 
-    const amounts = [];
-    await db.collection("orderHistory").find({"ownerNumber": "1"}).sort({"menu": 1}).forEach((r) => {
-        idx = amounts[idx] === undefined ? idx + 1 : idx;
-        amounts.splice(idx, 0, amounts[idx] ? amounts[idx] + 1 : 1);
-    });
-
-    idx = 0;
-    const sales = [];
-    const dayOfTheWeek = [];
-    await db.collection("orderHistory").find({"ownerNumber": "1"}).sort({"date": 1}).forEach((r) => {
-        if (idx !== 0) {
-            if (dayOfTheWeek[idx - 1] !== r.date) {
-                dayOfTheWeek[idx] = r.date;
-                sales[idx++] = Number(r.amount);
-            } else {
-                sales[idx - 1] += Number(r.amount);
-            }
-        } else {
-            dayOfTheWeek[idx] = r.date;
-            sales[idx++] = Number(r.amount);
-        }
-
-    });
-
-    result = {result, pizzaName: pizzaName, amounts: amounts, sales: sales, dayOfTheWeek: dayOfTheWeek};
-    return res.json(result);
-})
-
-app.post("/MenuReg", form_data.single("img"), (req, res) => {
-    db.collection("pizza").insertOne({...req.body, img: req.file}).then(() => {
-        console.log("데이터 삽입 성공");
-    });
-    return res.json(req.body);
-})
-
-/*app.post("/tempDummyInsert", form_data.any(), (req, res) => {
-    console.log(req.body);
-    db.collection("orderHistory").insertOne(req.body).then(() => {
-        console.log("데이터 삽입 성공");
-    });
-    return res.json(req.body);
-})*/
-
-app.get("/test", async (req, res) => {
-    let result = [];
-    await db.collection("pizza").find().forEach((r) => {
-        result.push({pizzaImg: r.img, pizzaName: r.name, pizzaPrice: r.priceL, pizzaOption: "테스트"});
-    })
-    //TODO 피자 컬렉션이 아니라 장바구니 컬렉션에서 조회하기
-    return res.json(result);
+    return res.json(count);
 })
 
 app.listen(port, hostname, () => {
