+++ src/components/pizza/Pizza.css
@@ -332,7 +332,7 @@
     -webkit-appearance: none;
     -moz-appearance: none;
     appearance: none;
-    width: 60px;
+    width: 100%;
     font-family: "Noto Sans KR Light",serif;
     font-size: 20px;
     color: #ccc;
