+++ src/components/pizza/Pizza.js
@@ -1,22 +1,10 @@
-import React, {useState, useEffect} from 'react';
+import React, {useState} from 'react';
 import "./Pizza.css";
 
-import icon_view from "./img/icon-view.png";
-import icon_basket from "./img/icon-basket.png";
-import pizza_test from "./img/master1.png";
-
-const PizzaDataUrl = "http://localhost:4000/api/loadPizzaData";
+import PizzaMenu from "./PizzaMenu";
 
 const PizzaPage = () => {
-    const [activeTab, setActiveTab] = useState(0);
-    const [pizzaData, setPizzaData] = useState([]);
-
-    useEffect(() => {
-      fetch(PizzaDataUrl)
-        .then(response => response.json())
-        .then(data => setPizzaData(data))
-        .catch(error => console.error(error));
-    }, []);
+    const [activeTab, setActiveTab] = useState("전체"); // 중단탭 피자 분류
 
     return(
         <div>
@@ -26,7 +14,7 @@ const PizzaPage = () => {
                         <div className="pc-pizzamenu">
                             <PizzaTopText></PizzaTopText>
                             <PizzaTab activeTab={activeTab} setActiveTab={setActiveTab}></PizzaTab>
-                            <PizzaMenu activeTab={activeTab} pizzaData={pizzaData}></PizzaMenu>
+                            <PizzaMenu activeTab={activeTab}></PizzaMenu>
                         </div>
                     </div>
                 </div>
@@ -55,8 +43,8 @@ function PizzaTab({activeTab, setActiveTab}) {
     }
 
     const tabList = ["전체", "장인피자", "달인피자", "명품피자"];
-    const handleTabClick = (index) => {
-        setActiveTab(index);
+    const handleTabClick = (type) => {
+        setActiveTab(type);
     }
 
     return(
@@ -66,150 +54,11 @@ function PizzaTab({activeTab, setActiveTab}) {
                     <tbody>
                         <tr className="column">
                             {tabList.map((tab, index) => (
-                                <td key={index} onClick={() => handleTabClick(index)} className={`tab-title ${activeTab===index ? "active" : "default"}`} style={tabStyle}>{tab}</td>
+                                <td key={index} onClick={() => handleTabClick(tab)} className={`tab-title ${activeTab===tab ? "active" : "default"}`} style={tabStyle}>{tab}</td>
                             ))}
                         </tr>
                     </tbody>
                 </table>
-            </div>
-        </div>
-    );
-}
-
-function PizzaMenu({activeTab, pizzaData}) {
-    const pageNum = Math.floor((pizzaData.length + 1) / 2);
-
-    const [selected, setSelected] = useState("2");
-    const handleSelectChange = (event) => {
-        setSelected(event.target.value);
-    }
-    useEffect(() => {
-        if(activeTab !== 0) {
-            setSelected("2");
-        }
-    }, [activeTab]);
-
-    const selectClassName = (activeTab !== 0 ? "alvolo-select-disable" : "alvolo-select-noline");
-
-    return(
-        <div className="pc-pizzamenu-menu">
-            <div className="pizzamenu-text-area">
-                <div className="pizzamenu-allergie-circle-area">
-                    <div className="pizzamenu-allergie-circle">i</div>
-                    <font color="#f9423a">재료성분</font>
-                </div>
-                <div className="pizzamenu-rank">
-                    <select id="selectActive" value={selected} disabled={activeTab !== 0 ? "disabled" : ""} className={selectClassName} style={{height:"68px"}} onChange={handleSelectChange}>
-                        <option value="2">신제품순</option>
-                        <option value="1">가격낮은순</option>
-                        <option value="0">가격높은순</option>
-                    </select>
-                </div>
-            </div>
-            <PizzaList pizzaData={pizzaData}></PizzaList>
-            <PizzaPagination pageNum={pageNum}></PizzaPagination>
-        </div>
-    );
-}
-
-function PizzaList({pizzaData}) {
-    return(
-        <div className="pc-pizzamenu-card-area">
-            {pizzaData.map((pizza, index) => (
-                <PizzaItem key={index} pizza={pizza}></PizzaItem>
-            ))}
-        </div>
-    );
-}
-
-function PizzaItem({pizza}) {
-    console.log(pizza);
-    const img = require(`${pizza.imgPath}`)
-    return(
-        <div className="pizzamenu-area-item">
-            <div className="carditem-web-container">
-                <div className="image-container">
-                    <img src={img} className="image" alt='#'></img>
-                </div>
-                <div className="item">
-                    <h5 className='item-name'>{pizza.name}</h5>
-                    <h6 className='item-description'>{pizza.description}</h6>
-                    <h5 className='pizzamenu-item-price'>
-                        <span>
-                            <span className='large'> L </span>
-                            {pizza.priceL.toLocaleString()}
-                        </span>
-                        <span>
-                            <span className='regular'> R </span>
-                            {pizza.priceR.toLocaleString()}
-                        </span>
-                    </h5>
-                    <div className='topping-box-container topping-box'>
-                        <div className="box">{pizza.topping1}</div>
-                    </div>
-                    <div className='topping-box-container topping-box'>
-                        <div className="box">{pizza.topping2}</div>
-                    </div>
-                    <div className='topping-box-container topping-box'>
-                        {pizza.topping3 !== "" && <div className="box">{pizza.topping3}</div>}
-                    </div>
-                </div>
-            </div>
-            <div className="click-active shadow-box">
-                <div className="view-button">
-                    <div className='inner-container'>
-                        <img src={icon_view} className="view-card" alt='#'></img>
-                        상세보기
-                    </div>
-                </div>
-                <div className="basket-button">
-                    <div className="inner-container">
-                        <img src={icon_basket} className="basket-card" alt='#'></img>
-                        장바구니
-                    </div>
-                </div>
-            </div>
-        </div>
-    );
-}
-
-function PizzaPagination({pageNum}) {
-    const maxPage = pageNum;
-    const minPage = 1;
-
-    const [activePage, setActivePage] = useState(minPage);
-    const pageItems = [];
-
-    const addPageNum = () => {
-        setActivePage(activePage + 1);
-    }
-
-    const subPageNum = () => {
-        setActivePage(activePage - 1);
-    }
-
-    const handleClickPage = (index) => {
-        setActivePage(index);
-    }
-
-    for(let i = 1; i <= pageNum; i++) {
-        pageItems.push(<li key={i} onClick={() => handleClickPage(i)} className={activePage === i ? "number active" : "number"}>{i}</li>)
-    }
-
-    return(
-        <div className="pizzamenu-pagination">
-            <div className="layout-pagination-whole">
-                <div className="layout-pagination el-pagination">
-                    <button type="button" onClick={() => subPageNum()} disabled={activePage <= minPage ? "disabled" : ""} className="btn-prev">
-                        <i className="el-icon el-icon-arrow-left"></i>
-                    </button>
-                    <ul className="el-pager">
-                        {pageItems}
-                    </ul>
-                    <button type="button" onClick={() => addPageNum()} disabled={activePage >= maxPage ? "disabled" : ""} className="btn-next">
-                        <i className="el-icon el-icon-arrow-right"></i>
-                    </button>
-                </div>
             </div>
         </div>
     );
